+++ src/pages/test/bottom-sheet/index.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 
 import useModal from '@/hooks/useModal';
 import useToast from '@/hooks/useToast';
-import BottomSheet from '@/components/common/Modal/BottomSheet';
+import BottomSheet from '@/components/common/Modal/Base/BottomSheet';
 
 const BOTTOM_SHEET_ID = 'bottomSheetId';
